dqs-orchestrator/pyproject.toml
[project]
name = "dqs-orchestrator"
version = "0.1.0"
description = "Data Quality Service — Python CLI orchestrator for spark-submit and email notifications"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2-binary>=2.9.11",
    "pyyaml>=6.0.3",
]
